backend/requirements.txt
flask
faiss-cpu
sentence-transformers
pandas
numpy
openai
